requirements.txt
asgiref==3.2.3
astroid==2.3.2
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
colorama==0.4.1
configparser==4.0.2
db.py==0.5.3
Django==3.0
django-filter==2.2.0
django-rest-auth==0.9.5
django-rest-authtoken==1.2.2
djangorestframework==3.10.3
Flask==1.1.1
Flask-Bcrypt==0.7.1
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.17.4
pandas==0.25.3
prettytable==0.7.2
psycopg2==2.8.4
pybars3==0.9.7
pycparser==2.19
pylint==2.4.3
PyMeta3==0.5.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.11
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
